fingerprint_detection_cnn/requirements.txt
streamlit
torch
torchvision
numpy
pandas
pillow
opencv-python
matplotlib